Create index.html
This is taken from Devine's example up on his github. Since this is going to be using React, not much other than metadata needs to go in this. All other parts of the app will be rendered at the <div> with id="root" using .js components. I changed the title, description, and commented out the theme-color for now.

I will be uploading placeholder icons for Apple menu and browser use.