Added more routing documentation
Documentation for all current routes is completed. I was unable to test the new /weather route. Frontend will have to mess around with it to see if its working because I can not tell if my attempted POST request from my browser's developers tools failed because the route isn't working, because I did something wrong with the request, or if its cros shenanigans.