Get Favorites Route
Added the /favorites route to backend/routes/weather.js to access a user's favorited beaches.
Added documentation for /favorites.
Modified /login to return the jwt and modified documentation to reflect this.
Removed code that attempted to set token cookie as it wasn't actually working.
Removed cookie-parser package as it is no longer used.
Revised /auth_test and jwtVerify() to reflect that backend is no longer dealing in cookies.